complete apart of Service, Controller of Author, Category, Book, Review

+++ src/main/java/BookInventoryManage/example/inventory/Modules/Book/BookService.java
@@ -1,17 +1,19 @@
 package BookInventoryManage.example.inventory.Modules.Book;
 
+import BookInventoryManage.example.inventory.Modules.Author.AuthorService;
 import BookInventoryManage.example.inventory.Modules.Book.DTO.CreateBookRequestDTO;
 import BookInventoryManage.example.inventory.Modules.Book.DTO.UpdateBookRequestDTO;
 import BookInventoryManage.example.inventory.Modules.Category.CategoryService;
+import BookInventoryManage.example.inventory.Modules.Databases.Entities.AuthorEntity;
 import BookInventoryManage.example.inventory.Modules.Databases.Entities.BookCategoryEntity;
 import BookInventoryManage.example.inventory.Modules.Databases.Entities.BookEntity;
 import BookInventoryManage.example.inventory.Modules.Databases.Entities.CategoryEntity;
 import BookInventoryManage.example.inventory.Modules.Databases.Repositories.BookCategoryRepository;
 import BookInventoryManage.example.inventory.Modules.Databases.Repositories.BookRepository;
-import BookInventoryManage.example.inventory.Modules.Databases.Repositories.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
@@ -28,12 +30,17 @@ public class BookService {
     CategoryService categoryService;
 
     @Autowired
+    AuthorService authorService;
+
+    @Autowired
     BookCategoryRepository bookCategoryRepository;
 
 
+    @Transactional
     public void addBook(CreateBookRequestDTO dto) {
         List<CategoryEntity> listCate = categoryService.getListCateByIds(dto.getCategoryIds());
-        BookEntity book = new BookEntity(dto);
+        AuthorEntity author = authorService.getAuthorByID(dto.getAuthor().getId());
+        BookEntity book = new BookEntity(dto, author);
         book = bookRepository.save(book);
         List<BookCategoryEntity> bookCategoryEntities = new ArrayList<>();
         for (CategoryEntity category : listCate) {
@@ -42,19 +49,16 @@ public class BookService {
         bookCategoryRepository.saveAll(bookCategoryEntities);
     }
 
-    //    nếu thay vì truyền só thì truyền kí tự có làm sao ?
-    public BookEntity getBookById(Integer Id) {
-        if (Id == null) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Category ID list must not be empty.");
-        }
-        Optional<BookEntity> OptBook = bookRepository.findById(Id);
-        if (OptBook.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found book !!");
-        else return OptBook.get();
-    }
-
+    @Transactional
     public void updateBook(Integer bookId, UpdateBookRequestDTO dto) {
         BookEntity book = getBookById(bookId);
+        if (dto.getTile() == null || dto.getTile().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Title must not be blank or whitespace only !!");
+        }
         book.setTitle(dto.getTile());
+        if (dto.getDescription() == null || dto.getDescription().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Description must not be blank or whitespace only !!");
+        }
         book.setDescription(dto.getDescription());
 
         bookCategoryRepository.deleteByBook(book);
@@ -68,16 +72,57 @@ public class BookService {
         bookCategoryRepository.saveAll(bookCategoryEntities);
     }
 
+//    @Transactional
+//    public void deleteBookById(Integer Id) {
+//        BookEntity book = getBookById(Id);
+//        bookRepository.delete(book);
+//        bookCategoryRepository.deleteByBook(book);
+//    }
+
     public List<BookEntity> listAllBooks() {
         List<BookEntity> books = bookRepository.findAll();
         if (books.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found any book !!");
         else return books;
     }
 
-    public void deleteBookById(Integer Id) {
-        BookEntity book = getBookById(Id);
-        bookRepository.delete(book);
-        bookCategoryRepository.deleteByBook(book);
+    //    search
+    public BookEntity getBookById(Integer Id) {
+        if (Id == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Book ID list must not be empty !!");
+        }
+        Optional<BookEntity> OptBook = bookRepository.findById(Id);
+        if (OptBook.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found book !!");
+        else return OptBook.get();
     }
 
+    public List<BookEntity> listBookByAuthorName(String authorName) {
+        if (authorName == null || authorName.trim().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Author's name must not be blank or whitespace only !!");
+        }
+        List<BookEntity> list = bookRepository.findByAuthor_NameContainingIgnoreCase(authorName);
+        if (list.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found any book of author's name " + authorName);
+        else return list;
+    }
+
+    public List<BookEntity> listBookByName(String name) {
+        if (name == null || name.trim().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Book's name must not be blank or whitespace only !!");
+        }
+        List<BookEntity> list = bookRepository.findByTitleContainingIgnoreCase(name);
+        if (list.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found any book by name " + name);
+        else return list;
+    }
+
+    public List<BookEntity> listBookByISBN(String isbn) {
+        if (isbn == null || isbn.trim().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ISBN must not be blank or whitespace only !!");
+        }
+        List<BookEntity> list = bookRepository.findByIsbnContainingIgnoreCase(isbn);
+        if (list.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found any book with ISBN = " + isbn);
+        }
+        return list;
+    }
 }
